backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.6.1
python-multipart==0.0.20
pillow==11.1.0
sqlalchemy==2.0.36
google-auth==2.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
email-validator==2.2.0
